% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_factor_functions.R
\name{create_age_factor}
\alias{create_age_factor}
\title{Function to Transform Age Factor}
\usage{
create_age_factor(x)
}
\arguments{
\item{x}{Name of analysis dataset}
}
\description{
This function transforms the AGE_YEARS variable into age_factor
}
\examples{
create_age_factor()
}
\keyword{age_factor}
