% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_factor_functions.R
\name{create_disability_factor}
\alias{create_disability_factor}
\title{This function transforms the DISABILITY_SHORT_NAME variable into disability_factor}
\usage{
create_disability_factor(x)
}
\arguments{
\item{x}{Name of analysis dataset}
}
\description{
This function transforms the DISABILITY_SHORT_NAME variable into disability_factor
}
\examples{
create_disability_factor()
}
\keyword{disability_factor}
