% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_factor_functions.R
\name{create_education_factor}
\alias{create_education_factor}
\title{Function to Transform Education Factor}
\usage{
create_education_factor(x)
}
\arguments{
\item{x}{Name of analysis dataset}
}
\description{
This function transform the raw EDUCATION_LEVEL variable into education_factor
}
\examples{
create_education_factor()
}
\keyword{education_factor}
