% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_factor_functions.R
\name{create_ethnicity_factor}
\alias{create_ethnicity_factor}
\title{Function to Transform Ethnicity Factor}
\usage{
create_ethnicity_factor(x)
}
\arguments{
\item{x}{Name of analysis dataset}
}
\description{
This function transforms the RACE_SHORT_NAME variable into ethnicity_factor
}
\examples{
create_ethnicity_factor()
}
\keyword{ethnicity_factor}
