% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_factor_functions.R
\name{create_job_series_factor}
\alias{create_job_series_factor}
\title{This function transforms the JOB_SERIES variable into job_series_factor}
\usage{
create_job_series_factor(x)
}
\arguments{
\item{x}{Name of analysis dataset}
}
\description{
This function transforms the JOB_SERIES variable into job_series_factor
}
\examples{
create_job_series_factor()
}
\keyword{job_series_factor}
