% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_factor_functions.R
\name{create_length_of_service_factor}
\alias{create_length_of_service_factor}
\title{Function to Transform Length of Service Factor}
\usage{
create_length_of_service_factor(x)
}
\arguments{
\item{x}{Name of analysis dataset}
}
\description{
This function transforms the LENGTH_OF_SERVICE_YEARS variable into length_of_service_factor
}
\examples{
create_length_of_service_factor()
}
\keyword{length_of_service_factor}
