% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_factor_functions.R
\name{create_location_factor}
\alias{create_location_factor}
\title{This function transforms the DUTY_STATION_STATE_CD variable into location_factor}
\usage{
create_location_factor(x)
}
\arguments{
\item{x}{Name of analysis dataset}
}
\description{
This function transforms the DUTY_STATION_STATE_CD variable into location_factor
}
\examples{
create_location_factor()
}
\keyword{job_series_factor}
