% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_factor_functions.R
\name{create_race_factor}
\alias{create_race_factor}
\title{Function to Transform Race Factor}
\usage{
create_race_factor(x)
}
\arguments{
\item{x}{Name of analysis dataset}
}
\description{
This function transforms the RACE_LONG_NAME variable into race_factor
}
\examples{
create_race_factor()
}
\keyword{race_factor}
