% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_factor_functions.R
\name{create_supervisor_status_factor}
\alias{create_supervisor_status_factor}
\title{This function transforms the SUPERVISOR_LONG_NAME variable into supervisor_status_factor}
\usage{
create_supervisor_status_factor(x)
}
\arguments{
\item{x}{Name of analysis dataset}
}
\description{
This function transforms the SUPERVISOR_LONG_NAME variable into supervisor_status_factor
}
\examples{
create_supervisor_status_factor()
}
\keyword{supervisor_status_factor}
