% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_factor_functions.R
\name{create_veteran_status_factor}
\alias{create_veteran_status_factor}
\title{Function to Transform Length of Service Factor}
\usage{
create_veteran_status_factor(x)
}
\arguments{
\item{x}{Name of analysis dataset}
}
\description{
This function transforms the VETERANS_STATUS_LONG_NAME variable into veteran_status_factor
}
\examples{
create_veteran_status_factor()
}
\keyword{veteran_status_factor}
