% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_factor_functions.R
\name{create_work_schedule_factor}
\alias{create_work_schedule_factor}
\title{This function transforms the WORK_SCHEDULE_LONG_NAME variable into work_schedule_factor}
\usage{
create_work_schedule_factor(x)
}
\arguments{
\item{x}{Name of analysis dataset}
}
\description{
This function transforms the WORK_SCHEDULE_LONG_NAME variable into work_schedule_factor
}
\examples{
create_work_schedule_factor()
}
\keyword{work_schedule_factor}
