% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_visualization_functions.R
\name{get_covariate_labels}
\alias{get_covariate_labels}
\title{Function to Get Covariate Labels}
\usage{
get_covariate_labels()
}
\description{
This function returns a vector with plain-text labels for covariates that are shown
in the Shiny dashboard
}
\examples{
get_covariate_labels()
I am only including the factor variables, so you'll note that pay is not included
This should include everything that's supposed to be in our full model
}
\keyword{covariate_labels}
