% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_visualization_functions.R
\name{get_covariates}
\alias{get_covariates}
\title{Function to Get Covariates}
\usage{
get_covariates()
}
\description{
This function returns a vector with the covariates that are reported on
}
\examples{
get_covariates()
}
\keyword{covariates}
