% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_visualization_functions.R
\name{get_government_wide_adjusted_means}
\alias{get_government_wide_adjusted_means}
\title{Function to Get Government-wide Regression Adjusted Means}
\usage{
get_government_wide_adjusted_means(models)
}
\arguments{
\item{models}{List of models to get the regression-adjusted means from}
}
\description{
This function outputs a data frame with government-wide regression adjusted means
}
\examples{
get_government_wide_adjusted_means()
}
\keyword{adjusted}
\keyword{estimated}
\keyword{marginal}
\keyword{means}
\keyword{means,}
\keyword{regression}
