% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_visualization_functions.R
\name{get_government_wide_means}
\alias{get_government_wide_means}
\title{Function to Get Government-wide Means}
\usage{
get_government_wide_means(x, question_codes, groups)
}
\arguments{
\item{x}{dataset with responses and weights to get means from}

\item{question_codes}{vector with column names to get statistics for}

\item{groups}{vector with quosures of covariate column names}
}
\description{
This function outputs a data frame with government-wide means, weighted means, SD, weighted SD, and number of responses
}
\examples{
get_government_wide_means(analysis_dataset, question_codes, groups)
}
\keyword{deviation}
\keyword{mean,}
\keyword{standard}
\keyword{weighted}
