% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_visualization_functions.R
\name{get_groups}
\alias{get_groups}
\title{Function to Get Covariates}
\usage{
get_groups()
}
\description{
This function returns a vector with the covariate groups that are reported on
}
\examples{
get_groups()
}
\keyword{groups}
