% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_visualization_functions.R
\name{get_var_map}
\alias{get_var_map}
\title{Function to Map Labels with Variable Names}
\usage{
get_var_map()
}
\description{
This function returns a map between variable names and labels for the Shiny
dashboard
}
\examples{
get_var_map()
}
\keyword{var_map}
