% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_functions.R
\name{getmode}
\alias{getmode}
\title{Function to Get Mode from Vector}
\usage{
getmode(v)
}
\arguments{
\item{v}{Vector to get mode from}
}
\description{
This function returns the mode from a vector
}
\examples{
getmode()
}
\keyword{mode}
