% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_functions.R
\name{impute_age_factor}
\alias{impute_age_factor}
\title{Function to Impute age_factor}
\usage{
impute_age_factor(x)
}
\arguments{
\item{x}{Name of analysis dataset}
}
\description{
This function imputes NA values for age_factor
}
\examples{
impute_age_factor()
}
\keyword{age_factor}
