% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_functions.R
\name{impute_education_factor}
\alias{impute_education_factor}
\title{Function to Impute Education Factor}
\usage{
impute_education_factor(x)
}
\arguments{
\item{x}{Name of analysis dataset}
}
\description{
This function imputes NA values for education_factor
}
\examples{
impute_education_factor()
}
\keyword{education}
