% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_functions.R
\name{impute_gender}
\alias{impute_gender}
\title{Function to Impute GENDER}
\usage{
impute_gender(x)
}
\arguments{
\item{x}{Name of analysis dataset}
}
\description{
This function imputes NA and '*' values for GENDER
}
\examples{
impute_gender()
}
\keyword{gender}
