% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_functions.R
\name{impute_length_of_service_factor}
\alias{impute_length_of_service_factor}
\title{Function to Impute length_of_service_factor}
\usage{
impute_length_of_service_factor(x)
}
\arguments{
\item{x}{Name of analysis dataset}
}
\description{
This function imputes NA values for length_of_service_factor
}
\examples{
impute_length_of_service_factor()
}
\keyword{length_of_service_factor}
