% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_functions.R
\name{impute_supervisor_status_factor}
\alias{impute_supervisor_status_factor}
\title{Function to Impute supervisor_status_factor}
\usage{
impute_supervisor_status_factor(x)
}
\arguments{
\item{x}{Name of analysis dataset}
}
\description{
This function imputes NA values for supervisor_status_factor
}
\examples{
impute_supervisor_status_factor()
}
\keyword{supervisor_status_factor}
