% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_functions.R
\name{impute_telework_eligibility_factor}
\alias{impute_telework_eligibility_factor}
\title{Function to Impute telework_eligibility_factor}
\usage{
impute_telework_eligibility_factor(x)
}
\arguments{
\item{x}{Name of analysis dataset}
}
\description{
This function imputes NA values for telework_eligibility_factor
}
\examples{
impute_telework_eligibility_factor()
}
\keyword{telework_eligibility_factor}
