% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_functions.R
\name{impute_total_annual_pay}
\alias{impute_total_annual_pay}
\title{Function to Impute TOTAL_ANNUAL_PAY}
\usage{
impute_total_annual_pay(x)
}
\arguments{
\item{x}{Name of analysis dataset}
}
\description{
This function imputes NA values for TOTAL_ANNUAL_PAY
}
\examples{
impute_total_annual_pay()
}
\keyword{TOTAL_ANNUAL_PAY}
