% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kclass.R
\name{Kclass}
\alias{Kclass}
\title{K-class estimator (kappa parametrization)}
\usage{
Kclass(k, A, Z, Y)
}
\arguments{
\item{k}{K-class kappa (numeric)}

\item{A}{exogenous variables (vector/matrix/data.frame)}

\item{Z}{included exogenous and endogenous regressors (vector/matrix/data.frame)}

\item{Y}{target variable (vector/matrix/data.frame)}
}
\value{
the K-class estimator with kappa parameter k
}
\description{
`Kclass` computes the K-class estimate for given kappa parameter k.
}
\details{
Contains several estimators. k=0: OLS, k=1: TSLS, k=Fuller_k(alpha) Fuller(alpha), k=LIML_k: LIML.
}
\examples{
### Just-identified example ###
# Generate data
n <- 500
A <- rnorm(n)
H <- rnorm(n)
X <- 0.8*A + H + rnorm(n)
Y <- 0.5*X + H + rnorm(n)
# Compute the OLS estimate
Kclass(k = 0, A = A, Z = X, Y = Y)
# Compute the TSLS estimate
Kclass(k = 1, A = A, Z = X, Y = Y)
# Compute the LIML estimate
Kclass(k = LIML_k(A,X,Y), A = A, Z = X, Y = Y)
# Compute the Fuller(4) estimate
Kclass(k = Fuller_k(4,A,X,Y), A = A, Z = X, Y = Y)

### Over-identified example ###
# Generate data
n  <- 500
A1 <- rnorm(n)
A2 <- rnorm(n)
A3 <- rnorm(n)
A  <- cbind(A1,A2,A3)
H  <- rnorm(n)
X  <- 0.8*A1 + 0.5*A2 + H + rnorm(n)
Y  <- 0.5*X + 0.2*A3 + H + rnorm(n)
# Compute the OLS estimate
Kclass(k = 0, A = A, Z = cbind(A3,X), Y = Y)
# Compute the TSLS estimate
Kclass(k = 1, A = A, Z = cbind(A3,X), Y = Y)
# Compute the LIML estimate
Kclass(k = LIML_k(A, X, Y, A_inc = A3), A = A, Z = cbind(A3,X), Y = Y)
# Compute the Fuller(4) estimate
Kclass(k = Fuller_k(4, A, X, Y, A_inc = A3), A = A, Z = cbind(A3,X), Y = Y)

}
