% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kclass.R
\name{LIML_k}
\alias{LIML_k}
\title{Limited information maximum likelihood (LIML) K-class kappa}
\usage{
LIML_k(A, X, Y, A_inc = NULL)
}
\arguments{
\item{A}{exogenous variables (vector/matrix/data.frame)}

\item{X}{included endogenous regressors (vector/matrix/data.frame)}

\item{Y}{target variable (vector/matrix/data.frame)}

\item{A_inc}{included exogenous regressors (vector/matrix/data.frame)}
}
\value{
the K-class kappa parameter corresponding to the LIML estimator
}
\description{
\code{LIML_k} compute the LIML K-class kappa parameter.
}
\examples{
### Just-identified example ###
# Generate data
n <- 500
A <- rnorm(n)
H <- rnorm(n)
X <- 0.8*A + H + rnorm(n)
Y <- 0.5*X + H + rnorm(n)
# Compute the LIML kappa parameter
LIML_k(A, X, Y)
# Compute the LIML estimate
Kclass(k = LIML_k(A,X,Y), A = A, Z = X, Y = Y)

### Over-identified example ###
# Generate data
n  <- 500
A1 <- rnorm(n)
A2 <- rnorm(n)
A3 <- rnorm(n)
A  <- cbind(A1,A2,A3)
H  <- rnorm(n)
X  <- 0.8*A1 + 0.5*A2 + H + rnorm(n)
Y  <- 0.5*X + 0.2*A3 + H + rnorm(n)
# Compute the LIML kappa parameter
LIML_k(A, X, Y, A_inc = A3)
# Compute the LIML estimate
Kclass(k = LIML_k(A, X, Y, A_inc = A3), A = A, Z = cbind(A3,X), Y = Y)

}
