% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_functions.R
\name{relevel_factors}
\alias{relevel_factors}
\title{Function to Relevel Factor}
\usage{
relevel_factors(x)
}
\arguments{
\item{x}{Name of analysis dataset}
}
\description{
This function relevels factor variables to set the base level, and removes unused level
}
\examples{
relevel_factors()
}
\keyword{relevel}
